move required constants to `shared` package
