+++ rasa/core/actions/action.py
@@ -23,9 +23,9 @@ from rasa.nlu.constants import (
     RESPONSE_SELECTOR_RESPONSES_KEY,
     RESPONSE_SELECTOR_PREDICTION_KEY,
     INTENT_RANKING_KEY,
-    INTENT_NAME_KEY,
     INTENT_RESPONSE_KEY,
 )
+from rasa.shared.nlu.constants import INTENT_NAME_KEY
 
 from rasa.shared.core.events import (
     UserUtteranceReverted,
