+++ rasa/core/training/story_reader/yaml_story_reader.py
@@ -14,7 +14,7 @@ from rasa.core.actions.action import RULE_SNIPPET_ACTION_NAME
 from rasa.shared.core.events import UserUttered, SlotSet, ActiveLoop
 from rasa.core.training.story_reader.story_reader import StoryReader
 from rasa.core.training.structures import StoryStep
-from rasa.nlu.constants import INTENT_NAME_KEY
+from rasa.shared.nlu.constants import INTENT_NAME_KEY
 import rasa.data
 
 logger = logging.getLogger(__name__)
