+++ rasa/nlu/constants.py
@@ -67,7 +67,6 @@ RESPONSE_IDENTIFIER_DELIMITER = "/"
 
 INTENT_RANKING_KEY = "intent_ranking"
 PREDICTED_CONFIDENCE_KEY = "confidence"
-INTENT_NAME_KEY = "name"
 
 FEATURE_TYPE_SENTENCE = "sentence"
 FEATURE_TYPE_SEQUENCE = "sequence"
