+++ tests/core/test_interpreter.py
@@ -6,7 +6,7 @@ from rasa.core.interpreter import (
     RasaNLUHttpInterpreter,
     RegexInterpreter,
 )
-from rasa.nlu.constants import INTENT_NAME_KEY
+from rasa.shared.nlu.constants import INTENT_NAME_KEY
 from rasa.utils.endpoints import EndpointConfig
 from tests.utilities import latest_request, json_of_latest_request
