+++ tests/core/training/story_reader/test_common_story_reader.py
@@ -14,7 +14,7 @@ from rasa.core.featurizers import (
     MaxHistoryTrackerFeaturizer,
     BinarySingleStateFeaturizer,
 )
-from rasa.nlu.constants import INTENT_NAME_KEY
+from rasa.shared.nlu.constants import INTENT_NAME_KEY
 
 
 @pytest.mark.parametrize(
