+++ tests/nlu/classifiers/test_fallback_classifier.py
@@ -11,12 +11,8 @@ from rasa.nlu.classifiers.fallback_classifier import (
     AMBIGUITY_THRESHOLD_KEY,
 )
 from rasa.nlu.training_data import Message
-from rasa.nlu.constants import (
-    INTENT_RANKING_KEY,
-    INTENT,
-    PREDICTED_CONFIDENCE_KEY,
-    INTENT_NAME_KEY,
-)
+from rasa.nlu.constants import INTENT_RANKING_KEY, INTENT, PREDICTED_CONFIDENCE_KEY
+from rasa.shared.nlu.constants import INTENT_NAME_KEY
 
 
 @pytest.mark.parametrize(
